Wrap token stream constructor in try-block
The token stream must rewind the internal iterator in order to perform
as expected while keeping the simplistic contract of the corresponding
interface. However, this makes it so the first token is tokenized within
the constructor. By moving the token stream construct to the try-catch
block of the program, whenever the first token is an unexpected token,
this is correctly caught.