Ensure all remaing strategies are valid options (#1)
When trying to resolve the (currently hypothetical) T_ADD token `+`, the
token will partially match the number strategy.

In the current setup, this makes the number strategy the top-most
remaining strategy left in the `$previous` array. Since only the
top-most remaining strategy is ever tried, the symbol strategy, that
will validate `+` on its own, never gets the chance to do so.

By reducing the `$previous` array until a token is found, the
performance is nigh equal to the previous situation, for all other
tokens, but allows a more comprehensive check for those tokens that need
it.