Split up TokenStream and simplify Optimizer
The token stream is now a combination of an iterator and a token
pointer, where the token pointer forces the signature of the `current`
method of the iterator.

By doing this, the parse context can still inherit from the tokenizer
namespace, yet the token optimizer can start implementing a shared
interface with the stateless tokenizer. This makes the optimization
layer of the tokenizer into a decorator, which makes the calling code
unaware of its existence.