Reduce cyclomatic complexity 11 -> 9
The cyclomatic complexity could be reduced by removing two if-conditions
that change nothing in the behavior of the token finder.

The first if allowed for an early exit, but this scenario is not
something to be optimized anyway.

The second if could be rewritten by using a short circuit using `??`.