Share project level IDE settings (#4)
* Move "tests" to "examples" to create room for unit tests

* Split up TokenStream and simplify Optimizer

The token stream is now a combination of an iterator and a token
pointer, where the token pointer forces the signature of the `current`
method of the iterator.

By doing this, the parse context can still inherit from the tokenizer
namespace, yet the token optimizer can start implementing a shared
interface with the stateless tokenizer. This makes the optimization
layer of the tokenizer into a decorator, which makes the calling code
unaware of its existence.

* Set up PHPUnit

* Cover Token

* Add Github Action for PHPUnit

* Replace Docker container with pre-made Github actions

* Previous actions do not mix: invoke manually

* Allow PHP 8

* Clean up action names

* Add unit test section

Add a section explaining how to perform unit tests and update references
to example files.

* Run unit tests on Ubuntu, Windows and macOS

To ensure compatibility between operating systems, the unit tests run on
Windows, macOS and Ubuntu.

* Rename jobs to contain runner

* Run PHPUnit using phpunit.xml.dist for compatibility with Windows CLI

* Do not configure coverage reporting

Let code coverage reports be configured locally through `phpunit.xml`
instead of `phpunit.xml.dist`. This allows for local development
configuration that is not shared with other developers - more
importantly not shared with CI configuration.

* Store IDE configuration in project repository

* Share IDE configurations

By sharing the IDE configurations, developers work with the same
standards, tools and should produce the same results between their work
environments.

* Add watcher for *.sym files

* Remove merge conflict

* Remove coverage/ from ignored docker context